refactor: 만다라트 생성폼을 MandalForm 컴포넌트로 분리
